requirements.txt
streamlit
nltk
openai
instructor
pydantic
python-dotenv
ratelimit
groq
anthropic